requirements.txt
discord.py==2.4.0
google-genai==0.3.0
google-cloud-secret-manager==2.20.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0
google-auth>=2.25.0
holidays>=0.40
